added terraform documentation, moved code into checkly package
