+++ checkly/resource_check_group_test.go
@@ -1,4 +1,4 @@
-package main
+package checkly
 
 import (
 	"regexp"
